move vol2col from fluid to phi
