+++ paddle/fluid/operators/math/CMakeLists.txt
@@ -45,7 +45,6 @@ endif()
 math_library(matrix_bit_code)
 
 math_library(unpooling)
-math_library(vol2col)
 math_library(prelu)
 math_library(bert_encoder_functor)
 math_library(tree2col DEPS math_function)
