+++ paddle/fluid/operators/math/vol2col_test.cc
@@ -12,10 +12,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-#include "paddle/fluid/operators/math/vol2col.h"
+#include "paddle/phi/kernels/funcs/vol2col.h"
 
 #include <gtest/gtest.h>
 
+#include "paddle/fluid/framework/tensor_util.h"
 #include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/place.h"
 
@@ -84,7 +85,7 @@ void testVol2col() {
                               output_width},
                              *place);
 
-  paddle::operators::math::Vol2ColFunctor<DeviceContext, float> vol2col;
+  phi::funcs::Vol2ColFunctor<DeviceContext, float> vol2col;
   vol2col(*context, input, dilations, strides, paddings, &output);
 
   float vol_2_col[] = {0, 1, 1, 2, 3, 4, 4, 5, 6, 7, 7, 8, 9, 10, 10, 11};
@@ -110,7 +111,7 @@ void testVol2col() {
     paddle::framework::TensorCopySync(input_tmp, *place, &input);
   }
 
-  paddle::operators::math::Col2VolFunctor<DeviceContext, float> col2vol;
+  phi::funcs::Col2VolFunctor<DeviceContext, float> col2vol;
   col2vol(*context, output, dilations, strides, paddings, &input);
 
   float* in_ptr;
@@ -201,7 +202,7 @@ void testVol2col<phi::GPUContext, paddle::platform::CUDAPlace>() {
                               output_width},
                              *place);
 
-  paddle::operators::math::Vol2ColFunctor<phi::GPUContext, float> vol2col;
+  phi::funcs::Vol2ColFunctor<phi::GPUContext, float> vol2col;
   vol2col(*context, input, dilations, strides, paddings, &output);
 
   float vol_2_col[] = {0, 1, 1, 2, 3, 4, 4, 5, 6, 7, 7, 8, 9, 10, 10, 11};
@@ -227,7 +228,7 @@ void testVol2col<phi::GPUContext, paddle::platform::CUDAPlace>() {
     paddle::framework::TensorCopySync(input_tmp, *place, &input);
   }
 
-  paddle::operators::math::Col2VolFunctor<phi::GPUContext, float> col2vol;
+  phi::funcs::Col2VolFunctor<phi::GPUContext, float> col2vol;
   col2vol(*context, output, dilations, strides, paddings, &input);
 
   float* in_ptr;
