+++ paddle/phi/kernels/funcs/CMakeLists.txt
@@ -16,6 +16,7 @@ math_library(pooling DEPS dense_tensor)
 math_library(segment_pooling)
 math_library(sequence2batch)
 math_library(matrix_solve DEPS dense_tensor eigen3 blas math_function)
+math_library(vol2col)
 
 cc_library(
   phi_data_layout_transform
