+++ paddle/phi/kernels/impl/conv_grad_kernel_impl.h
@@ -15,11 +15,11 @@
 #pragma once
 
 #include "paddle/fluid/operators/math/im2col.h"
-#include "paddle/fluid/operators/math/vol2col.h"
 #include "paddle/phi/kernels/cpu/conv_util.h"
 #include "paddle/phi/kernels/funcs/batch_norm_utils.h"
 #include "paddle/phi/kernels/funcs/blas/blas.h"
 #include "paddle/phi/kernels/funcs/math_function.h"
+#include "paddle/phi/kernels/funcs/vol2col.h"
 
 namespace phi {
 
@@ -147,7 +147,7 @@ void ConvGradKernel(const Context& dev_ctx,
     if (is_expand) {
       set_zero(dev_ctx, &transformed_input_grad, static_cast<T>(0));
     }
-    paddle::operators::math::Col2VolFunctor<Context, T> col2vol;
+    phi::funcs::Col2VolFunctor<Context, T> col2vol;
     paddle::operators::math::
         Col2ImFunctor<paddle::operators::math::ColFormat::kCFO, Context, T>
             col2im;
@@ -206,7 +206,7 @@ void ConvGradKernel(const Context& dev_ctx,
     paddle::operators::math::
         Im2ColFunctor<paddle::operators::math::ColFormat::kCFO, Context, T>
             im2col;
-    paddle::operators::math::Vol2ColFunctor<Context, T> vol2col;
+    phi::funcs::Vol2ColFunctor<Context, T> vol2col;
     for (int i = 0; i < batch_size; i++) {
       DenseTensor out_grad_batch =
           transformed_output_grad.Slice(i, i + 1).Resize(output_matrix_shape);
@@ -381,7 +381,7 @@ void ConvGradGradKernel(const Context& dev_ctx,
     if (is_expand) {
       set_zero(dev_ctx, &transformed_dX, static_cast<T>(0));
     }
-    paddle::operators::math::Col2VolFunctor<Context, T> col2vol;
+    phi::funcs::Col2VolFunctor<Context, T> col2vol;
     paddle::operators::math::
         Col2ImFunctor<paddle::operators::math::ColFormat::kCFO, Context, T>
             col2im;
@@ -431,7 +431,7 @@ void ConvGradGradKernel(const Context& dev_ctx,
     paddle::operators::math::
         Im2ColFunctor<paddle::operators::math::ColFormat::kCFO, Context, T>
             im2col;
-    paddle::operators::math::Vol2ColFunctor<Context, T> vol2col;
+    phi::funcs::Vol2ColFunctor<Context, T> vol2col;
     for (int i = 0; i < batch_size; ++i) {
       DenseTensor dy_batch =
           transformed_dY.Slice(i, i + 1).Resize(output_matrix_shape);
@@ -480,7 +480,7 @@ void ConvGradGradKernel(const Context& dev_ctx,
     paddle::operators::math::
         Im2ColFunctor<paddle::operators::math::ColFormat::kCFO, Context, T>
             im2col;
-    paddle::operators::math::Vol2ColFunctor<Context, T> vol2col;
+    phi::funcs::Vol2ColFunctor<Context, T> vol2col;
     for (int i = 0; i < batch_size; ++i) {
       DenseTensor ddy_batch =
           transformed_ddY.Slice(i, i + 1).Resize(output_matrix_shape);
